Move old /examples to /demo
