+++ src/gql/pages/demo/native-features/example-with-ssr.ts
